Fixed bug when CONSUME ORGANS presents non-capped spleen use
When spleen usage is not capped to the spleen limit (e.g. CONSUME SIM ORGANS 10 10 0 NOMEAT), it could previously end up with 10/10/15 spleen usage due to a bug where the diet attempts to "fill" the spleen cleansed by Sliders or Pickle Juices (even though the spleen usage was 0 and nothing was cleansed)

This fix adds a buffer to the spleen so that if any cleansers are consumed, there would actually be spleen to be cleansed. Any buffered spleen use is then removed from the end of the diet so that we do not consume past the user-defined limit after cleansing.